+++ packages/yak-swc/yak_swc/tests/fixture/css-prop-invalid/output.prod.tsx
@@ -1,3 +1,4 @@
 import { css, __yak_mergeCssProp } from "next-yak/internal";
-const Elem = ()=><div/>;
+const styles = /*#__PURE__*/ css();
+const Elem = ()=><div css="invalid"/>;
 const Elem2 = (props: any)=><div {...__yak_mergeCssProp({}, props)}/>;